Move test for SeamHttpRequest into its own file
